devex-1115: terraform for new cron to check migration progress
