+++ web-api/migration-cron-terraform/main/lambdas/reindex-status.js
@@ -1,8 +1,9 @@
-const axios = require('axios');
+const {
+  approvePendingJob,
+} = require('../../../../shared/admin-tools/circleci/approve-pending-job');
 const {
   isReindexComplete,
 } = require('../../../../shared/admin-tools/elasticsearch/check-reindex-complete');
-const { find } = require('lodash');
 
 exports.handler = async () => {
   const migrateFlag = process.env.MIGRATE_FLAG;
@@ -23,26 +24,5 @@ exports.handler = async () => {
   const apiToken = process.env.CIRCLE_MACHINE_USER_TOKEN;
   const workflowId = process.env.CIRCLE_WORKFLOW_ID;
 
-  const get_all_jobs = {
-    headers: { 'Circle-Token': apiToken },
-    method: 'GET',
-    url: `https://circleci.com/api/v2/workflow/${workflowId}/job`,
-  };
-
-  const allJobsInWorkflow = await axios.get(get_all_jobs.url, get_all_jobs);
-
-  const jobWithApprovalNeeded = find(
-    allJobsInWorkflow.data.items,
-    function (o) {
-      return o.approval_request_id !== undefined;
-    },
-  );
-
-  const approveJob = {
-    headers: { 'Circle-Token': apiToken },
-    method: 'POST',
-    url: `https://circleci.com/api/v2/workflow/${workflowId}/approve/${jobWithApprovalNeeded.approval_request_id}`,
-  };
-
-  await axios.post(approveJob.url, {}, approveJob);
+  await approvePendingJob({ apiToken, workflowId });
 };
